testutil: init mocking io by copying mbbsd/io.c to testutil/mock_io.c
We would like to test the functions requiring io.
Found that mocking mbbsd/io.c is the easiest method to work-around io.

This is to have the record that we directly copying mbbsd/io.c to testutil/mock_io.c
to be able to do git diff on testutil/mock_io.c in the future.